Fix: If a raft instance of the partition is migrated to the same node twice by frequent decommission action, it will delete itself by applying the ConfChange raft log.

+++ datanode/partition_raft.go
@@ -326,6 +326,10 @@ func (dp *DataPartition) addRaftNode(req *proto.DataPartitionDecommissionRequest
 
 // Delete a raft node.
 func (dp *DataPartition) removeRaftNode(req *proto.DataPartitionDecommissionRequest, index uint64) (isUpdated bool, err error) {
+	var canRemoveSelf bool
+	if canRemoveSelf, err = dp.canRemoveSelf(); err != nil {
+		return
+	}
 	peerIndex := -1
 	data, _ := json.Marshal(req)
 	isUpdated = false
@@ -354,7 +358,8 @@ func (dp *DataPartition) removeRaftNode(req *proto.DataPartitionDecommissionRequ
 		dp.config.Hosts = append(dp.config.Hosts[:hostIndex], dp.config.Hosts[hostIndex+1:]...)
 	}
 	dp.config.Peers = append(dp.config.Peers[:peerIndex], dp.config.Peers[peerIndex+1:]...)
-	if dp.config.NodeID == req.RemovePeer.ID {
+
+	if dp.config.NodeID == req.RemovePeer.ID && canRemoveSelf {
 		dp.raftPartition.Delete()
 		dp.Disk().space.DeletePartition(dp.partitionID)
 		isUpdated = false
