+++ metanode/partition_fsmop.go
@@ -110,6 +110,10 @@ func (mp *metaPartition) confAddNode(req *proto.
 
 func (mp *metaPartition) confRemoveNode(req *proto.MetaPartitionDecommissionRequest,
 	index uint64) (updated bool, err error) {
+	var canRemoveSelf bool
+	if canRemoveSelf, err = mp.canRemoveSelf(); err != nil {
+		return
+	}
 	peerIndex := -1
 	data, _ := json.Marshal(req)
 	log.LogInfof("Start RemoveRaftNode  PartitionID(%v) nodeID(%v)  do RaftLog (%v) ",
@@ -127,7 +131,7 @@ func (mp *metaPartition) confRemoveNode(req *proto.MetaPartitionDecommissionRequ
 		return
 	}
 	mp.config.Peers = append(mp.config.Peers[:peerIndex], mp.config.Peers[peerIndex+1:]...)
-	if mp.config.NodeId == req.RemovePeer.ID {
+	if mp.config.NodeId == req.RemovePeer.ID && canRemoveSelf {
 		mp.Stop()
 		mp.DeleteRaft()
 		mp.manager.deletePartition(mp.GetBaseConfig().PartitionId)
